+++ xarray/tests/test_backends_datatree.py
@@ -265,6 +265,19 @@ class DatatreeIOBase:
             assert_equal(original_dt, roundtrip_dt)
             assert_identical(expected_dt, roundtrip_dt)
 
+    def test_compute_false(self, tmpdir, simple_datatree):
+        filepath = tmpdir / "test.nc"
+        original_dt = simple_datatree.chunk()
+        result = original_dt.to_netcdf(filepath, engine=self.engine, compute=False)
+
+        with open_datatree(filepath, engine=self.engine) as in_progress_dt:
+            assert in_progress_dt.isomorphic(original_dt)
+            assert not in_progress_dt.equals(original_dt)
+
+        result.compute()
+        with open_datatree(filepath, engine=self.engine) as written_dt:
+            assert_identical(written_dt, original_dt)
+
 
 @requires_netCDF4
 class TestNetCDF4DatatreeIO(DatatreeIOBase):
@@ -575,6 +588,9 @@ class TestZarrDatatreeIO:
         with open_datatree(filepath, engine="zarr") as roundtrip_dt:
             assert_equal(original_dt, roundtrip_dt)
 
+    @pytest.mark.filterwarnings(
+        "ignore:Numcodecs codecs are not in the Zarr version 3 specification"
+    )
     def test_zarr_encoding(self, tmpdir, simple_datatree, zarr_format):
         filepath = str(tmpdir / "test.zarr")
         original_dt = simple_datatree
@@ -601,11 +617,10 @@ class TestZarrDatatreeIO:
 
             enc["/not/a/group"] = {"foo": "bar"}  # type: ignore[dict-item]
             with pytest.raises(ValueError, match="unexpected encoding group.*"):
-                original_dt.to_zarr(
-                    filepath, encoding=enc, engine="zarr", zarr_format=zarr_format
-                )
+                original_dt.to_zarr(filepath, encoding=enc, zarr_format=zarr_format)
 
     @pytest.mark.xfail(reason="upstream zarr read-only changes have broken this test")
+    @pytest.mark.filterwarnings("ignore:Duplicate name")
     def test_to_zarr_zip_store(self, tmpdir, simple_datatree, zarr_format):
         from zarr.storage import ZipStore
 
@@ -653,7 +668,9 @@ class TestZarrDatatreeIO:
 
         storepath = tmp_path / "test.zarr"
         original_dt = simple_datatree.chunk()
-        original_dt.to_zarr(str(storepath), compute=False, zarr_format=zarr_format)
+        result = original_dt.to_zarr(
+            str(storepath), compute=False, zarr_format=zarr_format
+        )
 
         def assert_expected_zarr_files_exist(
             arr_dir: Path,
@@ -723,6 +740,14 @@ class TestZarrDatatreeIO:
                     is_scalar=not bool(var.dims),
                     zarr_format=zarr_format,
                 )
+
+        with open_datatree(str(storepath), engine="zarr") as in_progress_dt:
+            assert in_progress_dt.isomorphic(original_dt)
+            assert not in_progress_dt.equals(original_dt)
+
+        result.compute()
+        with open_datatree(str(storepath), engine="zarr") as written_dt:
+            assert_identical(written_dt, original_dt)
 
     def test_to_zarr_inherited_coords(self, tmpdir, zarr_format):
         original_dt = DataTree.from_dict(